+++ CHANGELOG.md
@@ -10,7 +10,7 @@ This project uses [Semantic Versioning][sv].
 - Add module `PwnedPasswords` to encapsulate details of the API that serves the
   known breach data.
 - Create the `Password` class to encapsulate details of the k-anonymity model.
-- Create the `API` class to handle networking details.
+- Create the `APIv2` class to handle networking details.
 
 ## [0.2.0][0.2.0] — 2020-01-28